backend/requirements.txt
# Backend Dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
websockets>=12.0
httpx>=0.26.0
pydantic>=2.5.3
motor>=3.3.2
pymongo>=4.6.1
groq>=0.4.2
google-generativeai>=0.3.2
deepgram-sdk>=3.0.0
numpy>=1.26.0
langsmith
langgraph
langchain
langchain-groq
pytest>=7.4.4
pytest-asyncio>=0.23.3
